+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
-- Yet more clean field `Debug`
+- Yet more clean field & register `Debug`
 
 ## [v0.33.2] - 2024-05-07